Revert "Remove the deprecated `otel/metric/global` pkg (#3829)"
This reverts commit 60f7d42d1eedae6381f1d6524374b57a274e1639.
