+++ metric/internal/global/instruments_test.go
