+++ metric/internal/global/meter.go
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global // import "go.opentelemetry.io/otel/internal/global"
+package global // import "go.opentelemetry.io/otel/metric/internal/global"
 
 import (
 	"container/list"
 	"sync"
 	"sync/atomic"
 
+	oGlob "go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/metric/instrument"
@@ -35,6 +36,11 @@ type meterProvider struct {
 	meters map[il]*meter
 
 	delegate metric.MeterProvider
+}
+
+type il struct {
+	name    string
+	version string
 }
 
 // setDelegate configures p to delegate all MeterProvider functionality to
@@ -335,7 +341,7 @@ func (c *registration) setDelegate(m metric.Meter) {
 
 	reg, err := m.RegisterCallback(c.function, insts...)
 	if err != nil {
-		GetErrorHandler().Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 	}
 
 	c.unreg = reg.Unregister
