+++ metric/internal/global/meter_types_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global // import "go.opentelemetry.io/otel/internal/global"
+package global // import "go.opentelemetry.io/otel/metric/internal/global"
 
 import (
 	"context"
